- doc tweak  - preparation for reusing instruction from existing stored files

+++ grails-app/services/org/objectrepository/instruction/WorkflowInitiateService.groovy
@@ -71,7 +71,7 @@ class WorkflowInitiateService extends WorkflowJob {
      */
     private void uploadInstruction(String na) {
         Instruction.findAllByNa(na)?.each {
-            if (!taskValidationService.hasFileSet(it)) {
+            if (it.fileSet[0] != '.' && !taskValidationService.hasFileSet(it)) {
                 it.delete() // We have become a lie... no fileset at all.
             } else if (it.task.name == 'UploadFiles') {
                 runMethod(it)
